[Refact](inverted index) refact inverted_index_query for range predicate
